fix(produtos): order price/cost history by timestamp; enforce int timestamp (Copilot)
Copilot review on #148:
- History modals queried historicoDePrecos/historicoDeCusto with limit(100) and
  NO orderBy, so Firestore returned an arbitrary subset and the client-side sort
  ran after the cap — with >100 records the modal could miss the newest. Add
  orderByField('timestamp', 'desc') before the limit so the cap keeps the newest
  rows; the client sort stays as a defensive tiebreak.
- historicoPrecoSchema/historicoCustoSchema typed timestamp as z.number(), but
  the wire value is an ms-epoch int (old rules: `d.timestamp is int`, and our
  writes use Date.now()). Tighten to z.number().int().

+++ packages/schemas/src/historicos.ts
@@ -20,7 +20,7 @@ export const historicoPrecoSchema = z
     listaDePrecoHistoricoOuterRef: z.string().min(1),
     valorOriginal: z.number().nullable().default(null),
     valorFinal: z.number().nullable().default(null),
-    timestamp: z.number().nullable().default(null),
+    timestamp: z.number().int().nullable().default(null),
   })
   .passthrough();
 
@@ -34,7 +34,7 @@ export type HistoricoPreco = z.infer<typeof historicoPrecoSchema>;
 export const historicoCustoSchema = z
   .object({
     valor: z.number().min(0),
-    timestamp: z.number().nullable().default(null),
+    timestamp: z.number().int().nullable().default(null),
   })
   .passthrough();
